Added separate hook
