+++ application/ui/src/features/annotator/tools/bounding-box-tool/bounding-box-tool.component.tsx
@@ -3,24 +3,18 @@
 
 import { useZoom } from '../../../../components/zoom/zoom.provider';
 import type { Label } from '../../../../constants/shared-types';
-import { useAnnotationActions } from '../../../../shared/annotator/annotation-actions-provider.component';
 import { useAnnotator } from '../../../../shared/annotator/annotator-provider.component';
-import { useSelectedAnnotations } from '../../../../shared/annotator/select-annotation-provider.component';
 import type { Rect } from '../../../../shared/types';
 import { DrawingBox } from '../drawing-box-tool/drawing-box.component';
+import { useAddAndSelectAnnotations } from '../use-add-and-select-annotations.hook';
 
 export const BoundingBoxTool = () => {
     const { scale: zoom } = useZoom();
-    const { addAnnotations } = useAnnotationActions();
-    const { setSelectedAnnotations } = useSelectedAnnotations();
+    const { addAndSelectAnnotations } = useAddAndSelectAnnotations();
     const { mediaItem, image, selectedLabel } = useAnnotator();
 
     const handleComplete = (shapes: Rect[], labels: Label[]): string[] => {
-        const newIds = addAnnotations(shapes, labels);
-
-        setSelectedAnnotations(new Set(newIds));
-
-        return newIds;
+        return addAndSelectAnnotations(shapes, labels);
     };
 
     return (
