+++ application/ui/src/features/annotator/tools/polygon-tool/polygon-tool.component.tsx
@@ -7,12 +7,11 @@ import { isPointOverPoint, isPolygonValid } from '@geti/smart-tools/utils';
 import { isEmpty } from 'lodash-es';
 
 import { useZoom } from '../../../../components/zoom/zoom.provider';
-import { useAnnotationActions } from '../../../../shared/annotator/annotation-actions-provider.component';
 import { useAnnotator } from '../../../../shared/annotator/annotator-provider.component';
-import { useSelectedAnnotations } from '../../../../shared/annotator/select-annotation-provider.component';
 import { Point } from '../../../../shared/types';
 import { usePolygonConfig } from '../hooks/use-polygon-config.hook';
 import { SvgToolCanvas } from '../svg-tool-canvas.component';
+import { useAddAndSelectAnnotations } from '../use-add-and-select-annotations.hook';
 import { PolygonDraw } from './polygon-draw.component';
 import {
     drawingStyles,
@@ -30,8 +29,7 @@ import classes from './polygon-tool.module.scss';
 
 export const PolygonTool = () => {
     const { scale: zoom } = useZoom();
-    const { addAnnotations } = useAnnotationActions();
-    const { setSelectedAnnotations } = useSelectedAnnotations();
+    const { addAndSelectAnnotations } = useAddAndSelectAnnotations();
     const { image, selectedLabel } = useAnnotator();
 
     const ref = useRef<SVGRectElement>({} as SVGRectElement);
@@ -122,12 +120,10 @@ export const PolygonTool = () => {
                 startTransition(async () => {
                     const optimizedPolygon = await optimizePolygonOrSegments(polygon);
 
-                    const newIds = addAnnotations(
+                    addAndSelectAnnotations(
                         [{ type: 'polygon', points: optimizedPolygon.points }],
                         selectedLabel ? [selectedLabel] : []
                     );
-
-                    setSelectedAnnotations(new Set(newIds));
                 });
 
                 resetTool();
